refactor: remove reimplimented code

+++ src/account_multisig.ts
@@ -6,12 +6,10 @@ import { Account, SignAndSendTransactionOptions } from './account';
 import { Connection } from './connection';
 import { parseNearAmount } from './utils/format';
 import { PublicKey } from './utils/key_pair';
-import { Action, addKey, deleteKey, deployContract, fullAccessKey, functionCall, functionCallAccessKey, signTransaction } from './transaction';
+import { Action, addKey, deleteKey, deployContract, fullAccessKey, functionCall, functionCallAccessKey } from './transaction';
 import { FinalExecutionOutcome, TypedError } from './providers';
 import { fetchJson } from './utils/web';
-import { AccessKeyView, FunctionCallPermissionView } from './providers/provider';
-import { KeyPair } from './utils/key_pair';
-import { baseDecode } from 'borsh';
+import { FunctionCallPermissionView } from './providers/provider';
 
 export const MULTISIG_STORAGE_KEY = '__multisigRequest';
 export const MULTISIG_ALLOWANCE = new BN(parseNearAmount('1'));
@@ -293,40 +291,35 @@ export class Account2FA extends AccountMultisig {
         }
     }
 
-    async disableWithFAK({ contractBytes, cleanupContractBytes }: { fullAccessKey: KeyPair, contractBytes: Uint8Array, cleanupContractBytes?: Uint8Array }) {
-        const publicKey = await this.connection.signer.getPublicKey(this.accountId, this.connection.networkId);
-        const accessKey = await this.connection.provider.query<AccessKeyView>({
-            request_type: 'view_access_key',
-            account_id: this.accountId,
-            public_key: publicKey.toString(),
-            finality: 'optimistic'
-        });
-        const block = await this.connection.provider.block({ finality: 'final' });
-        const blockHash = block.header.hash;
+    async disableWithFAK({ contractBytes, cleanupContractBytes }: { contractBytes: Uint8Array, cleanupContractBytes?: Uint8Array }) {
+        let cleanupActions = [];
+        const keyConversionActions = await this.get2faDisableKeyConversionActions();
+        if(cleanupContractBytes) {
+            await this.deleteAllRequests().catch(e => {
+                if(new RegExp(MultisigDeleteRequestRejectionError.REQUEST_COOLDOWN_ERROR).test(e?.kind?.ExecutionError)) {
+                    return e;
+                }
+                throw e;
+            });
+            cleanupActions = await this.get2faDisableCleanupActions(cleanupContractBytes);
+        }
 
-        if(accessKey.permission !== 'FullAccess') {
+        const actions = [
+            ...cleanupActions,
+            ...keyConversionActions,
+            deployContract(contractBytes)
+        ];
+
+        const accessKeyInfo = await this.findAccessKey(this.accountId, actions);
+
+        if(accessKeyInfo?.accessKey?.permission !== 'FullAccess') {
             throw new TypedError(`No full access key found in keystore. Unable to bypass multisig`, 'NoFAKFound');
         }
 
-        const nonce = ++accessKey.nonce;
-        const actions = [
-            ...(cleanupContractBytes ? await this.get2faDisableCleanupActions(cleanupContractBytes) : []),
-            ...(await this.get2faDisableKeyConversionActions()),
-            deployContract(contractBytes),
-        ]
-        const [, signedTx] = await signTransaction(
-            this.accountId, nonce, actions, baseDecode(blockHash), this.connection.signer, this.accountId, this.connection.networkId
-        );
-        return this.connection.provider.sendTransaction(signedTx)
+        return this.signAndSendTransactionWithAccount(this.accountId, actions);
     }
 
     async get2faDisableCleanupActions(cleanupContractBytes: Uint8Array) {
-        await this.deleteAllRequests().catch(e => {
-            if(new RegExp(MultisigDeleteRequestRejectionError.REQUEST_COOLDOWN_ERROR).test(e?.kind?.ExecutionError)) {
-                return e;
-            }
-            throw e;
-        });
         const currentAccountState: { key: Buffer, value: Buffer }[] = await this.viewState('').catch(error => {
             const cause = error.cause && error.cause.name;
             if (cause == 'NO_CONTRACT_CODE') {
@@ -363,6 +356,12 @@ export class Account2FA extends AccountMultisig {
     }
 
     async disable(contractBytes: Uint8Array, cleanupContractBytes: Uint8Array) {
+        await this.deleteAllRequests().catch(e => {
+            if(new RegExp(MultisigDeleteRequestRejectionError.REQUEST_COOLDOWN_ERROR).test(e?.kind?.ExecutionError)) {
+                return e;
+            }
+            throw e;
+        });
         const actions = [
             ...(await this.get2faDisableCleanupActions(cleanupContractBytes)),
             ...(await this.get2faDisableKeyConversionActions()),
